test(db): update similar search tests to remove Lucene dependencies
